# openapi.json
{"openapi":"3.1.0","info":{"title":"Unified Social Media API","description":"User suggestions + Post/Reel recommendations in one service","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/clear-embed-cache":{"post":{"summary":"Clear Embed Cache","description":"Manually flush the in-process embedding cache.","operationId":"clear_embed_cache_admin_clear_embed_cache_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/suggest/{user_id}":{"get":{"summary":"Suggest","operationId":"suggest_suggest__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"description":"Number of user suggestions to return","default":10,"title":"Limit"},"description":"Number of user suggestions to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recommend/{user_id}":{"get":{"summary":"Recommend","operationId":"recommend_recommend__user_id__get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"string","title":"User Id"}},{"name":"top_n","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"description":"Max posts / reels to return","default":20,"title":"Top N"},"description":"Max posts / reels to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"CommentItem":{"properties":{"id":{"type":"string","title":"Id"},"username":{"type":"string","title":"Username"},"avatar":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Avatar"},"post":{"type":"string","title":"Post"},"parent":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parent"},"content":{"type":"string","title":"Content"},"created_at":{"type":"string","title":"Created At"}},"type":"object","required":["id","username","avatar","post","parent","content","created_at"],"title":"CommentItem"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MediaItem":{"properties":{"id":{"type":"string","title":"Id"},"file":{"type":"string","title":"File"},"media_type":{"type":"string","title":"Media Type"}},"type":"object","required":["id","file","media_type"],"title":"MediaItem"},"PostDetail":{"properties":{"id":{"type":"string","title":"Id"},"user_id":{"type":"string","title":"User Id"},"username":{"type":"string","title":"Username"},"avatar":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Avatar"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"media":{"items":{"$ref":"#/components/schemas/MediaItem"},"type":"array","title":"Media"},"categories_detail":{"items":{},"type":"array","title":"Categories Detail"},"shared_post":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Shared Post"},"shared_post_details":{"anyOf":[{"$ref":"#/components/schemas/SharedPostDetails"},{"type":"null"}]},"reactions_count":{"type":"integer","title":"Reactions Count"},"like_count":{"type":"integer","title":"Like Count"},"reaction_types":{"items":{"type":"string"},"type":"array","title":"Reaction Types"},"current_user_reaction":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Current User Reaction"},"is_followed":{"type":"boolean","title":"Is Followed"},"comments_count":{"type":"integer","title":"Comments Count"},"comments":{"items":{"$ref":"#/components/schemas/CommentItem"},"type":"array","title":"Comments"},"views_count":{"type":"integer","title":"Views Count"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"},"final_score":{"type":"number","title":"Final Score"},"content_score":{"type":"number","title":"Content Score"},"trending_score":{"type":"number","title":"Trending Score"},"random_score":{"type":"number","title":"Random Score"}},"type":"object","required":["id","user_id","username","avatar","content","media","categories_detail","shared_post","shared_post_details","reactions_count","like_count","reaction_types","current_user_reaction","is_followed","comments_count","comments","views_count","created_at","updated_at","final_score","content_score","trending_score","random_score"],"title":"PostDetail"},"RecommendationResponse":{"properties":{"user_id":{"type":"string","title":"User Id"},"total_posts":{"type":"integer","title":"Total Posts"},"top_n":{"type":"integer","title":"Top N"},"posts":{"items":{"$ref":"#/components/schemas/PostDetail"},"type":"array","title":"Posts"}},"type":"object","required":["user_id","total_posts","top_n","posts"],"title":"RecommendationResponse"},"SharedPostDetails":{"properties":{"id":{"type":"string","title":"Id"},"username":{"type":"string","title":"Username"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"avatar":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Avatar"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"created_at":{"type":"string","title":"Created At"},"media":{"items":{"$ref":"#/components/schemas/MediaItem"},"type":"array","title":"Media"}},"type":"object","required":["id","username","full_name","avatar","content","created_at","media"],"title":"SharedPostDetails"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}